move time list components to new TimeLists folder
